Split traceback utilities that depend on Rich into cli subpackage
