+++ modal/__main__.py
@@ -1,7 +1,8 @@
 # Copyright Modal Labs 2022
 import sys
 
-from ._traceback import highlight_modal_deprecation_warnings, reduce_traceback_to_user_code, setup_rich_traceback
+from ._traceback import reduce_traceback_to_user_code
+from .cli._traceback import highlight_modal_deprecation_warnings, setup_rich_traceback
 from .cli.entry_point import entrypoint_cli
 from .cli.import_refs import _CliUserExecutionError
 from .config import config
